+++ app-frontend/src/app/components/common/listItem/index.js
@@ -1,37 +1,23 @@
 import _ from 'lodash';
 import tpl from './index.html';
 
-class ListItemController {
-    $onInit() {
-        const rx = /^#(?:[0-9a-f]{3}){1,2}$/i;
-        const color = this.color;
-        if (color && color.match(rx)) {
-            this.color = color;
-        } else {
-            this.color = 'black';
-        }
-    }
-}
-
 const component = {
     bindings: {
-        id: '<',
-        color: '<',
-        title: '<',
-        subtitle: '<',
-        date: '<',
-        getImage: '&?',
-        onImageClick: '&?',
-        selected: '<',
-        onSelect: '&?'
+        id: '<'
     },
     templateUrl: tpl,
-    controller: ListItemController.name,
-    transclude: true
+    transclude: {
+        selector: '?itemSelector',
+        title: 'itemTitle',
+        subtitle: '?itemSubtitle',
+        date: '?itemDate',
+        preview: '?itemPreview',
+        actions: '?itemActions',
+        statistics: '?itemStatistics'
+    }
 };
 
 export default angular
     .module('components.common.listItem', [])
-    .controller(ListItemController.name, ListItemController)
     .component('rfListItem', component)
     .name;