+++ app-frontend/src/app/components/common/listItemWidgets/index.js
@@ -1,0 +1,7 @@
+import listItemSelector from './listItemSelector';
+import listItemActions from './listItemActions';
+
+export default [
+    listItemSelector,
+    listItemActions
+];