+++ app-frontend/src/app/components/common/listItemWidgets/listItemActions/index.js
@@ -1,0 +1,13 @@
+import tpl from './index.html';
+
+const component = {
+    bindings: {
+        actions: '<'
+    },
+    templateUrl: tpl
+};
+
+export default angular
+    .module('components.common.listItemWidgets.listItemActions', [])
+    .component('rfListItemActions', component)
+    .name;