+++ app-frontend/src/app/components/exports/index.js
@@ -1,0 +1,5 @@
+import exportStatus from './exportStatus';
+
+export default [
+    exportStatus
+];