+++ app-frontend/src/app/components/lab/templateItem/templateItem.module.js
@@ -8,6 +8,7 @@ const TemplateItemComponent = {
         templateData: '<',
         onTemplateDelete: '&',
         onShareClick: '&',
+        onEditClick: '&',
         hideActions: '<'
     }
 };