+++ app-frontend/src/app/components/pages/project/layers/index.js
@@ -4,7 +4,7 @@ import {Set, Map} from 'immutable';
 
 class ProjectLayersPageController {
     constructor(
-        $rootScope, $state,
+        $rootScope, $state, $q,
         projectService, paginationService, modalService, authService, mapService
     ) {
         'ngInject';
@@ -42,8 +42,7 @@ class ProjectLayersPageController {
     }
 
     fetchPage(page = this.$state.params.page || 1) {
-        let layerList = [];
-        this.itemList = [];
+        this.layerList = [];
         this.layerActions = {};
         const currentQuery = this.projectService.getProjectLayers(
             this.project.id,
@@ -52,8 +51,8 @@ class ProjectLayersPageController {
                 page: page - 1
             }
         ).then((paginatedResponse) => {
-            layerList = paginatedResponse.results;
-            layerList.forEach((layer) => {
+            this.layerList = paginatedResponse.results;
+            this.layerList.forEach((layer) => {
                 layer.subtext = '';
                 if (layer.id === this.project.defaultLayerId) {
                     layer.subtext += 'Default layer';
@@ -62,13 +61,10 @@ class ProjectLayersPageController {
                     layer.subtext += layer.subtext.length ? ', Smart layer' : 'Smart Layer';
                 }
             });
-            const defaultLayer = layerList.find(l => l.id === this.project.defaultLayerId);
-            this.layerActions = layerList.map(
+            const defaultLayer = this.layerList.find(l => l.id === this.project.defaultLayerId);
+            this.layerActions = this.layerList.map(
                 (l) => this.addLayerActions(l, defaultLayer === l)
             );
-            this.itemList = layerList.map(layer => {
-                return this.createItemInfo(layer);
-            });
             this.pagination = this.paginationService.buildPagination(paginatedResponse);
             this.paginationService.updatePageParam(page);
             if (this.currentQuery === currentQuery) {
@@ -91,6 +87,33 @@ class ProjectLayersPageController {
         if (!this.permissions.includes('Edit')) {
             return [];
         }
+        const alertAoiAction = {
+            icon: 'icon-warning color-danger',
+            name: 'Edit AOI',
+            tooltip: 'No AOI defined for this layer',
+            callback: () => this.goToAoiDef(layer.id),
+            menu: false
+        };
+        const editAoiAction = {
+            name: 'Edit AOI',
+            callback: () => this.goToAoiDef(layer.id),
+            menu: true
+        };
+        const previewAction = {
+            icons: [
+                {
+                    icon: 'icon-eye',
+                    isActive: () => this.visible.has(layer.id)
+                }, {
+                    icon: 'icon-eye-off',
+                    isActive: () => !this.visible.has(layer.id)
+                }
+            ],
+            name: 'Preview',
+            tooltip: 'Show/hide on map',
+            callback: () => this.onHide(layer.id),
+            menu: false
+        };
         const editAction = {
             name: 'Edit',
             callback: () => this.$state.go(
@@ -153,24 +176,29 @@ class ProjectLayersPageController {
         };
         const deleteAction = {
             name: 'Delete',
-            callback: () => this.deleteProjectLayer(layer),
+            callback: () => this.deleteProjectLayers(layer),
             menu: true
         };
         const section = {
             section: true
         };
 
-        const unimplementedActions = [publishAction, exportAction, settingsAction];
-        const layerActions = [editAction, createAnalysisAction, importAction, browseAction];
+        // TODO: add implement these when the views are finished:
+        // const unimplementedActions = [publishAction, exportAction, settingsAction];
+        const layerActions = [
+            previewAction, editAction, editAoiAction, createAnalysisAction,
+            importAction, browseAction
+        ];
 
         return [
+            ...!_.get(layer, 'geometry.type') ? [alertAoiAction] : [],
             ...layerActions,
             ...!isDefaultLayer ? [setDefaultAction, deleteAction] : []
         ];
     }
 
     allVisibleSelected() {
-        let layerSet = new Set(this.itemList.map(l => l.id));
+        let layerSet = new Set(this.layerList.map(l => l.id));
         return layerSet.intersect(this.selected.keySeq()).size === layerSet.size;
     }
 
@@ -179,7 +207,7 @@ class ProjectLayersPageController {
             this.selected = this.selected.clear();
         } else {
             this.selected = this.selected.merge(
-                new Map(this.itemList.map(i => [i.id, i]))
+                new Map(this.layerList.map(i => [i.id, i]))
             );
         }
         this.updateSelectText();
@@ -187,9 +215,9 @@ class ProjectLayersPageController {
 
     updateSelectText() {
         if (this.allVisibleSelected()) {
-            this.selectText = 'Clear selected';
+            this.selectText = `Clear selected (${this.selected.size})`;
         } else {
-            this.selectText = 'Select listed';
+            this.selectText = `Select all listed (${this.selected.size})`;
         }
     }
 
@@ -197,7 +225,7 @@ class ProjectLayersPageController {
         if (this.selected.has(id)) {
             this.selected = this.selected.delete(id);
         } else {
-            const layer = this.itemList.find(l => l.id === id);
+            const layer = this.layerList.find(l => l.id === id);
             this.selected = this.selected.set(id, layer);
         }
     }
@@ -213,10 +241,6 @@ class ProjectLayersPageController {
             this.visible = this.visible.add(id);
         }
         this.syncMapLayersToVisible();
-    }
-
-    isVisible(layerId) {
-        return this.visible.has(layerId);
     }
 
     setProjectDefaultLayer(layer) {
@@ -235,27 +259,58 @@ class ProjectLayersPageController {
         this.$state.go('project.create-analysis', {layers});
     }
 
-    deleteProjectLayer(layer) {
-        const modal = this.modalService.open({
-            component: 'rfFeedbackModal',
-            resolve: {
-                title: () => 'Really delete layer?',
-                subtitle: () => 'Deleting layers cannot be undone',
-                content: () =>
-                    '<h2>Do you wish to continue?</h2>'
-                    + '<p>Future attempts to access this '
-                    + 'layer or associated annotations, tiles, and scenes will fail.',
-                feedbackIconType: () => 'danger',
-                feedbackIcon: () => 'icon-warning',
-                feedbackBtnType: () => 'btn-danger',
-                feedbackBtnText: () => 'Delete layer',
-                cancelText: () => 'Cancel'
-            }
-        });
-        modal.result.then(() => {
-            this.projectService.deleteProjectLayer(this.project.id, layer.id).then(() => {
+    deleteProjectLayers(layers) {
+        let modal;
+        let ids = layers.length ? layers.map(l => l.id) : [layers.id];
+        if (ids.length > 1) {
+            modal = this.modalService.open({
+                component: 'rfFeedbackModal',
+                resolve: {
+                    title: () => `Really delete ${ids.length} layers?`,
+                    subtitle: () => 'Deleting layers cannot be undone',
+                    content: () =>
+                        '<h2>Do you wish to continue?</h2>'
+                        + '<p>Future attempts to access these '
+                        + 'layers or associated annotations, tiles, and scenes will fail.',
+                    feedbackIconType: () => 'danger',
+                    feedbackIcon: () => 'icon-warning',
+                    feedbackBtnType: () => 'btn-danger',
+                    feedbackBtnText: () => 'Delete layers',
+                    cancelText: () => 'Cancel'
+                }
+            }).result;
+        } else {
+            modal = this.modalService.open({
+                component: 'rfFeedbackModal',
+                resolve: {
+                    title: () => 'Really delete layer?',
+                    subtitle: () => 'Deleting layers cannot be undone',
+                    content: () =>
+                        '<h2>Do you wish to continue?</h2>'
+                        + '<p>Future attempts to access this '
+                        + 'layer or associated annotations, tiles, and scenes will fail.',
+                    feedbackIconType: () => 'danger',
+                    feedbackIcon: () => 'icon-warning',
+                    feedbackBtnType: () => 'btn-danger',
+                    feedbackBtnText: () => 'Delete layer',
+                    cancelText: () => 'Cancel'
+                }
+            }).result;
+        }
+        modal.then(() => {
+            const promises = ids.map(
+                id => this.projectService.deleteProjectLayer(this.project.id, id)
+            );
+            this.$q.all(promises).then(() => {
+                this.visible = this.visible.subtract(this.selected.keySeq());
+                this.selected = new Map();
+            }).catch(e => {
+                this.$log.error(e);
+            }).finally(() => {
                 this.fetchPage();
             });
+        }).catch(() => {
+            // modal closed
         });
     }
 
@@ -265,7 +320,7 @@ class ProjectLayersPageController {
     }
 
     showPageLayers() {
-        this.visible = this.visible.union(this.itemList.map(l => l.id));
+        this.visible = this.visible.union(this.layerList.map(l => l.id));
         this.syncMapLayersToVisible();
     }