+++ app-frontend/src/app/components/projects/index.js
@@ -1,17 +1,17 @@
-import layerItem from './layerItem';
 import layerStats from './layerStats';
 import projectLayerSecondaryNavbar from './projectLayerSecondaryNavbar';
 import selectedActionsBar from './selectedActionsBar';
 import navbar from './navbar';
 import projectSettingsNavbar from './projectSettingsNavbar';
 import aoiDrawToolbar from './aoiDrawToolbar';
+import analysisEditModal from './analysisEditModal';
 
 export default [
-    layerItem,
     projectSettingsNavbar,
     layerStats,
     projectLayerSecondaryNavbar,
     selectedActionsBar,
     navbar,
-    aoiDrawToolbar
+    aoiDrawToolbar,
+    analysisEditModal
 ];