+++ app-frontend/src/app/index.components.js
@@ -3,6 +3,7 @@ import pages from './components/pages';
 import projectComponents from './components/projects';
 import sceneComponents from './components/scenes';
 import commonComponents from './components/common';
+import exportComponents from './components/exports';
 export default angular.module('index.components', [
     //admin components
     require('./components/admin/editableLogo/editableLogo.js').default.name,
@@ -93,7 +94,7 @@ export default angular.module('index.components', [
     require('./components/filters/sliderFilter/sliderFilter.module.js').default.name,
     require('./components/filters/tagFilter/tagFilter.module.js').default.name,
     require('./components/filters/ingestedFilter/ingestedFilter.module.js').default.name,
-
+    ...exportComponents,
 
     // common components (no domain)
     require('./components/common/navBar/navBar.module.js').default.name,